# qlife-latest/qlife-server/openapi.yaml
openapi: 3.0.3
info:
  title: QLife API
  description: |
    REST API for the QLife mental-health platform. Supports three primary roles:
    - **USER** – a registered end-user seeking mental-health support
    - **PROFESSIONAL** – a verified mental-health professional
    - **ADMIN** – a platform administrator

    All protected endpoints require a Bearer JWT obtained from the Cognito-backed
    authentication flow. The token is passed in the `Authorization` header as
    `Bearer <token>`.

    All successful responses wrap their payload in a top-level `data` key.
    Error responses use `{ "error": { "code": "...", "message": "..." } }`.
  version: 1.0.0
  contact:
    name: QLife Engineering
servers:
  - url: https://api.qlife.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /assessments:
    get:
      tags: [Assessments]
      operationId: listAssessments
      summary: List assessments
      description: |
        Returns assessments for the authenticated account filtered by optional
        status. USERs see their own assessments (self-initiated or assigned by
        a professional). PROFESSIONALs see assessments they assigned to clients.
        Requires role **USER** or **PROFESSIONAL**.
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by assessment status (e.g. PENDING, COMPLETED)
          schema:
            type: string
            example: "PENDING"
      responses:
        '200':
          description: Assessment list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      assessments:
                        type: array
                        items:
                          $ref: '#/components/schemas/Assessment'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
              example:
                data:
                  assessments:
                    - id: "asmt-0000-0000-0000-000000000001"
                      status: "PENDING"
                      instrumentSlug: "phq-9"
                      instrumentName: "PHQ-9"
                      dueAt: null
                      createdAt: "2025-06-05T12:00:00.000Z"
                  pagination:
                    page: 1
                    pageSize: 20
                    total: 1
                    hasMore: false
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Assessments]
      operationId: createAssessment
      summary: Create a self-assessment
      description: |
        Creates a new self-initiated assessment for the authenticated account
        using the specified instrument slug. Available to any authenticated user
        (USER or PROFESSIONAL).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssessmentRequest'
            example:
              instrumentSlug: "phq-9"
      responses:
        '200':
          description: Newly created assessment
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      assessment:
                        $ref: '#/components/schemas/Assessment'
              example:
                data:
                  assessment:
                    id: "asmt-0000-0000-0000-000000000002"
                    status: "PENDING"
                    instrumentSlug: "phq-9"
                    instrumentName: "PHQ-9"
                    dueAt: null
                    createdAt: "2025-06-17T09:00:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /assessments/assign/{careRelationshipId}:
    post:
      tags: [Assessments]
      operationId: assignAssessment
      summary: Assign assessments to a client (Professional)
      description: |
        Allows a PROFESSIONAL to assign one or more instrument assessments to a
        client identified by a care-relationship ID. An optional due date can be
        set. Requires role **PROFESSIONAL**.
      parameters:
        - name: careRelationshipId
          in: path
          required: true
          description: Care-relationship UUID linking professional to client
          schema:
            type: string
            format: uuid
            example: "cr-000000-0000-0000-0000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignAssessmentRequest'
            example:
              instrumentSlugs: ["phq-9", "gad-7"]
              dueAt: "2025-07-01T00:00:00.000Z"
      responses:
        '200':
          description: List of created assessment records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      assessments:
                        type: array
                        items:
                          $ref: '#/components/schemas/Assessment'
              example:
                data:
                  assessments:
                    - id: "asmt-0000-0000-0000-000000000003"
                      status: "PENDING"
                      instrumentSlug: "phq-9"
                      instrumentName: "PHQ-9"
                      dueAt: "2025-07-01T00:00:00.000Z"
                      createdAt: "2025-06-17T09:05:00.000Z"
                    - id: "asmt-0000-0000-0000-000000000004"
                      status: "PENDING"
                      instrumentSlug: "gad-7"
                      instrumentName: "GAD-7"
                      dueAt: "2025-07-01T00:00:00.000Z"
                      createdAt: "2025-06-17T09:05:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /assessments/{id}/answers:
    post:
      tags: [Assessments]
      operationId: submitAssessmentAnswers
      summary: Submit answers for an assessment
      description: |
        Submits answers for a specific assessment. Each answer maps a
        `questionId` to a `selectedOptionId`. When all required questions are
        answered the assessment status transitions to COMPLETED and a scoring
        band is computed. Requires the authenticated account to own the
        assessment.
      parameters:
        - name: id
          in: path
          required: true
          description: Assessment UUID
          schema:
            type: string
            format: uuid
            example: "asmt-0000-0000-0000-000000000002"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswersRequest'
            example:
              answers:
                - questionId: "q-0000-0000-0000-000000000001"
                  selectedOptionId: "opt-0000-0000-0000-000000000003"
                - questionId: "q-0000-0000-0000-000000000002"
                  selectedOptionId: "opt-0000-0000-0000-000000000007"
              note: ""
      responses:
        '200':
          description: Assessment result after submission
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      assessment:
                        $ref: '#/components/schemas/AssessmentDetail'
              example:
                data:
                  assessment:
                    id: "asmt-0000-0000-0000-000000000002"
                    status: "COMPLETED"
                    instrumentSlug: "phq-9"
                    instrumentName: "PHQ-9"
                    score: 12
                    answers:
                      - questionId: "q-0000-0000-0000-000000000001"
                        selectedOptionId: "opt-0000-0000-0000-000000000003"
                    scoringBand:
                      label: "Moderate"
                      minScore: 10
                      maxScore: 14
                    dueAt: null
                    createdAt: "2025-06-17T09:00:00.000Z"
                    completedAt: "2025-06-17T09:12:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------- Clients -------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by Amazon Cognito. Pass as Authorization: Bearer <token>.
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "professionalProfileId must be a UUID"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Unauthorized"
    Forbidden:
      description: Authenticated account lacks the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "FORBIDDEN"
              message: "Forbidden"
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error description
              example: "professionalProfileId must be a UUID"
    Pagination:
      type: object
      description: Pagination metadata included in list responses
      required: [page, pageSize, total, hasMore]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        hasMore:
          type: boolean
          example: true
...
    Assessment:
      type: object
      description: Assessment summary
      properties:
        id:
          type: string
          format: uuid
          example: "asmt-0000-0000-0000-000000000001"
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, EXPIRED]
          example: "PENDING"
        instrumentSlug:
          type: string
          example: "phq-9"
        instrumentName:
          type: string
          example: "PHQ-9"
        dueAt:
          type: string
          format: date-time
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
          example: "2025-06-05T12:00:00.000Z"
    AssessmentDetail:
      allOf:
        - $ref: '#/components/schemas/Assessment'
        - type: object
          properties:
            score:
              type: number
              nullable: true
              example: 12
            answers:
              type: array
              items:
                type: object
                properties:
                  questionId:
                    type: string
                    format: uuid
                    example: "q-0000-0000-0000-000000000001"
                  selectedOptionId:
                    type: string
                    format: uuid
                    example: "opt-0000-0000-0000-000000000003"
            scoringBand:
              nullable: true
              type: object
              properties:
                label:
                  type: string
                  example: "Moderate"
                minScore:
                  type: number
                  example: 10
                maxScore:
                  type: number
                  example: 14
            completedAt:
              type: string
              format: date-time
              nullable: true
              example: "2025-06-17T09:12:00.000Z"
    CreateAssessmentRequest:
      type: object
      required: [instrumentSlug]
      properties:
        instrumentSlug:
          type: string
          minLength: 1
          example: "phq-9"
    AssignAssessmentRequest:
      type: object
      required: [instrumentSlugs]
      properties:
        instrumentSlugs:
          type: array
          items:
            type: string
          minItems: 1
          example: ["phq-9", "gad-7"]
        dueAt:
          type: string
          format: date-time
          example: "2025-07-01T00:00:00.000Z"
    SubmitAnswersRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          minItems: 1
          items:
            type: object
            required: [questionId, selectedOptionId]
            properties:
              questionId:
                type: string
                format: uuid
                example: "q-0000-0000-0000-000000000001"
              selectedOptionId:
                type: string
                format: uuid
                example: "opt-0000-0000-0000-000000000003"
        note:
          type: string
          example: ""

    # ---- Clients -----------------------------------------------------------
...